+++ examples/cuda-flake-parts/flake.nix
@@ -35,12 +35,8 @@
           buildType = "Release";
           compilerCppStd = "20";
           stdenv = pkgs.cudaPackages_13_2.backendStdenv;
-          platformToolRequires = {
-            cmake = pkgs.cmake.version;
-          };
           devShell = {
             tools = {
-              inherit (pkgs) cmake;
               inherit (pkgs.cudaPackages_13_2)
                 cuda_nvcc
                 cuda_cccl
@@ -59,11 +55,21 @@
               GALLIUM_DRIVER = "d3d12";
             };
           };
-          buildEnv = [
+          runEnv = [
             {
               name = "LD_LIBRARY_PATH";
               op = "+=(path)";
               value = "/usr/lib/wsl/lib";
+            }
+            {
+              name = "MESA_D3D12_DEFAULT_ADAPTER_NAME";
+              op = "=";
+              value = "NVIDIA";
+            }
+            {
+              name = "GALLIUM_DRIVER";
+              op = "=";
+              value = "d3d12";
             }
           ];
           remotes.local = {
@@ -78,21 +84,6 @@
             config.conan.outputs.devShell
           ];
         };
-
-        checks.test = pkgs.runCommandWith
-          {
-            name = "cuda-flake-parts-test-conan-create";
-            inherit (config.conan) stdenv;
-            derivationArgs = { inherit (config.conan.outputs.devShell) buildInputs nativeBuildInputs; };
-          }
-          ''
-            (
-            set -x
-            ${config.conan.outputs.devShell.shellHook}
-            conan create ${config.conan.info.configRoot} -tf="" --build=missing 2>&1 | grep "example/0.0.1"
-            touch $out
-            )
-          '';
       };
     };
   #  }