+++ examples/standalone-eval-conan-config/flake.nix
@@ -59,16 +59,7 @@
               modules = [
                 ({ pkgs, config, ... }: {
                   buildType = "Release";
-
                   compilerCppStd = "17";
-
-                  platformToolRequires = {
-                    cmake = pkgs.cmake.version;
-                  };
-
-                  devShell = {
-                    tools = { inherit (pkgs) cmake; };
-                  };
 
                   remotes.local = {
                     url = "./repo";